test(nextcloud): refactor test file structure
