+++ packages/nextcloud/test/api/webdav/csrf_client_test.dart
@@ -1,0 +1,20 @@
+import 'package:http_client_conformance_tests/http_client_conformance_tests.dart';
+import 'package:nextcloud/nextcloud.dart';
+import 'package:nextcloud/src/api/webdav/webdav.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group(
+    'Client conformance tests',
+    () {
+      testAll(
+        () => WebDavCSRFClient(NextcloudClient(Uri())),
+        canReceiveSetCookieHeaders: true,
+        canSendCookieHeaders: true,
+      );
+    },
+    onPlatform: const {
+      'browser': [Skip()],
+    },
+  );
+}
